+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Removed
 
+- 通貨転送取引において決済方法タイプ:PointをAccountへ自動変換する互換性維持対応を削除
+
 ### Fixed
 
 ### Security